Fix lint/format and extract MemberRow to its own file
- Remove redundant (trigger as HTMLElement) cast in GroupDetailView.spec.tsx
  line ~504 — screen.getByTestId already returns HTMLElement
- Extract MemberRow component and MemberRowProps interface into co-located
  MemberRow.tsx per CLAUDE.md row/list-item extraction rule; import it in
  GroupDetailView.tsx and remove now-unused DropdownMenu imports
- Rebase onto main to resolve merge conflicts with PR #211
  (kept this PR's more complete MemberRow: admin-visible menu, Remove from
  group, inline confirmation; merged both sets of copy keys alphabetically)
